# backend/src/docs/openapi/api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Web Booking Ticket Cinema Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for my application. \n \n Some useful links: \n - [Project repository](https://github.com/twilight2368/web-ticket-booking-cinema-project)"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "Successfully logged in"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/auth/admin-register": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Admin registered successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Admin username already exists"
          }
        }
      }
    },
    "/auth/admin-login": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "Admin logged in successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Admin not found"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
  }
}